dc_mrfld: fix 32bit code compatibility
Tracked-On: https://jira01.devtools.intel.com/browse/AW-4083
Reviewed-on: https://android.intel.com/565165
Tested-by: Bel Aj Ali, HabibX <habibx.bel.aj.ali@intel.com>

+++ drivers/external_drivers/intel_media/graphics/rgx/services/3rdparty/intel_drm/dc_mrfld.c
@@ -1522,7 +1522,7 @@ static PVRSRV_ERROR DC_MRFLD_ContextConfigureCheck(
 
 		/*copy the context from userspace*/
 		err = copy_from_user(psSurfCustom,
-				(void *)(pasSurfAttrib[i].ui64Custom),
+				(void *)(long)(pasSurfAttrib[i].ui64Custom),
 				sizeof(DC_MRFLD_SURF_CUSTOM));
 		if (err) {
 			DRM_ERROR("Failed to copy plane context\n");
